Implement live viewing over the network
This adds two new options, which allow the user to plug the IR camera
into a headless system, and view the live camera output from another.

To test over localhost:

	./ircam -d /dev/video4 -l &
	./ircam -c 127.0.0.1

It's uncompressed for now, because the lavc library has the baked in
assumption you're encoding to and decoding from files: later I'll
refactor it to work with buffers in both directions.