Autodetect cameras, add scaffolding for more models
Implement camera autodetection, so it is no longer necessary to pass the
device path on the command line.

Because it avoids redoing the same work later, go ahead and refactor the
core to support multiple camera models. This is very basic and obviously
will need more work in the long run.